WIP: add test
Also: fix tyConToFgTyCon; add FgType_Unit
